requirements.txt
streamlit
langchain
langchain-experimental
langchain-openai
python-dotenv
openai
pandas
tabulate